backend/app/services/email_service.py
"""
Email service for sending and managing emails.

This module handles email delivery via SendGrid, email logging,
and webhook processing for email events.
"""

from typing import List, Optional, Dict, Any
from app.models.email_log import EmailLog
from app.schemas.email_schema import EmailSendRequest, EmailResponse
from app.config import settings
from sendgrid import SendGridAPIClient
from sendgrid.helpers.mail import Mail
from app.db.mongodb import get_database

class EmailService:
    """Service for email sending and management."""
    
    def __init__(self, db):
        """Initialize email service with database and SendGrid."""
        self.db = db
        self.emails_collection = db.email_logs
        self.sg = SendGridAPIClient(api_key=settings.SENDGRID_API_KEY)
        self.from_email = settings.SENDGRID_FROM_EMAIL
        self.from_name = settings.SENDGRID_FROM_NAME
    
    async def send_email(self, email_data: EmailSendRequest) -> EmailResponse:
        """Send an email via SendGrid."""
        pass
    
    async def get_email(self, email_id: str) -> Optional[EmailResponse]:
        """Get an email by ID."""
        pass
    
    async def list_emails(
        self,
        page: int = 1,
        limit: int = 10,
        client_id: Optional[str] = None,
        status: Optional[str] = None
    ) -> List[EmailResponse]:
        """List emails with pagination and filtering."""
        pass
    
    async def log_email(
        self,
        task_id: str,
        client_id: str,
        to_email: str,
        subject: str,
        body: str
    ) -> EmailLog:
        """Log an email before sending."""
        pass
    
    async def update_email_status(
        self,
        email_id: str,
        status: str,
        sendgrid_message_id: Optional[str] = None,
        error_message: Optional[str] = None
    ) -> bool:
        """Update email status after sending or webhook events."""
        pass
    
    async def process_webhook_event(self, event_data: Dict[str, Any]) -> bool:
        """Process SendGrid webhook events (opens, clicks, bounces)."""
        pass
