backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
openai==1.3.8
python-sendgrid==6.10.0
pytest==7.4.3
pytest-asyncio==0.23.1
httpx==0.25.2
